# src/main/java/com/probestack/forgesphere/api/ComplianceRulesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgesphere.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import com.probestack.forgesphere.model.AssetType;
import com.probestack.forgesphere.model.ComplianceRulesResponse;
import com.probestack.forgesphere.model.CreateComplianceRuleRequest;
import com.probestack.forgesphere.model.CreateComplianceRuleResponse;
import com.probestack.forgesphere.model.ErrorResponse;
import com.probestack.forgesphere.model.RuleCategory;
import com.probestack.forgesphere.model.RuleStatus;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-13T05:38:50.494838724Z[GMT]")@Validated
@Tag(name = "Compliance Rules", description = "APIs for creating and fetching compliance rules.")
public interface ComplianceRulesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /governance/v1/compliance-rules : Add compliance rule
     * Adds a new compliance rule.  ruleId is generated by Java service and should not be sent by UI. 
     *
     * @param createComplianceRuleRequest  (required)
     * @return Compliance rule created successfully. (status code 201)
     *         or Invalid request. (status code 400)
     *         or Duplicate rule. (status code 409)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "createComplianceRule",
        summary = "Add compliance rule",
        description = "Adds a new compliance rule.  ruleId is generated by Java service and should not be sent by UI. ",
        tags = { "Compliance Rules" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Compliance rule created successfully.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreateComplianceRuleResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Duplicate rule.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/governance/v1/compliance-rules",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<CreateComplianceRuleResponse> createComplianceRule(
        @Parameter(name = "CreateComplianceRuleRequest", description = "", required = true) @Valid @RequestBody CreateComplianceRuleRequest createComplianceRuleRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"ruleName\" : \"Documentation required for all endpoints\", \"ruleId\" : \"CR13\", \"message\" : \"Compliance rule created successfully.\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /governance/v1/compliance-rules : Get compliance rules
     * Fetch compliance rules for UI display.
     *
     * @param projectName Project name used by UI context. (optional)
     * @param resourceType Resource type selected in the UI. (optional)
     * @param resourceName Resource name selected in the UI. (optional)
     * @param status Rule status filter. (optional)
     * @param ruleCategory Rule category filter. (optional)
     * @return Compliance rules fetched successfully. (status code 200)
     *         or Invalid request. (status code 400)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "getComplianceRules",
        summary = "Get compliance rules",
        description = "Fetch compliance rules for UI display.",
        tags = { "Compliance Rules" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Compliance rules fetched successfully.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComplianceRulesResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/governance/v1/compliance-rules",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ComplianceRulesResponse> getComplianceRules(
        @Parameter(name = "projectName", description = "Project name used by UI context.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "projectName", required = false) String projectName,
        @Parameter(name = "resourceType", description = "Resource type selected in the UI.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "resourceType", required = false) AssetType resourceType,
        @Parameter(name = "resourceName", description = "Resource name selected in the UI.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "resourceName", required = false) String resourceName,
        @Parameter(name = "status", description = "Rule status filter.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "status", required = false) RuleStatus status,
        @Parameter(name = "ruleCategory", description = "Rule category filter.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "ruleCategory", required = false) RuleCategory ruleCategory
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"resourceName\" : \"Remit2Any\", \"rules\" : [ { \"updatedBy\" : \"updatedBy\", \"displayOrder\" : 0, \"icon\" : \"icon\", \"updatedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"mandatory\" : true, \"ruleDescription\" : \"ruleDescription\", \"enabled\" : true, \"createdBy\" : \"createdBy\", \"ruleName\" : \"Documentation required for all endpoints\", \"ruleOwner\" : \"Governance Team\", \"ruleId\" : \"CR13\", \"createDate\" : \"2000-01-23T04:56:07.000+00:00\" }, { \"updatedBy\" : \"updatedBy\", \"displayOrder\" : 0, \"icon\" : \"icon\", \"updatedDate\" : \"2000-01-23T04:56:07.000+00:00\", \"mandatory\" : true, \"ruleDescription\" : \"ruleDescription\", \"enabled\" : true, \"createdBy\" : \"createdBy\", \"ruleName\" : \"Documentation required for all endpoints\", \"ruleOwner\" : \"Governance Team\", \"ruleId\" : \"CR13\", \"createDate\" : \"2000-01-23T04:56:07.000+00:00\" } ], \"projectName\" : \"Payment Gateway\", \"activeRules\" : 29, \"totalRules\" : 30 }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}