# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    API specification for probestack-compliance-service.

    This service manages governance compliance rules and supports compliance scanning
    for microservices, Apigee proxies, and Kong resources.

    ## Microservice

    Microservice Name: probestack-compliance-service

    ## MongoDB Collections

    ### governance_compliance_rules

    Stores compliance rule definitions used by the UI and scan engine.

    Primary Fields:
    - ruleId: Unique rule identifier generated by Java service. Example: CR13
    - ruleName: Compliance rule display name
    - ruleDescription: Detailed rule description
    - ruleType: PRE_DEFINED or CUSTOM
    - ruleOwner: Owner team or person
    - category: DOCUMENTATION, LOGGING, SECURITY, PERFORMANCE, HEALTH, GENERAL
    - severity: REQUIRED, MANDATORY, RECOMMENDED, OPTIONAL
    - enabled: Whether the rule is enabled
    - mandatory: Whether the rule is mandatory
    - displayOrder: UI display order
    - icon: UI icon name
    - status: complaint or non-compliant
    - createDate: Created timestamp
    - createdBy: User who created the rule
    - updatedDate: Last updated timestamp
    - updatedBy: User who last updated the rule

    Recommended Indexes:
    - Unique index on ruleId
    - Compound index on status, category, ruleType

    ### governance_compliance_scans

    Stores submitted, running, completed, and failed compliance scan executions.

    Primary Fields:
...
    - projectName: Project name
...
servers:
...
paths:
  /governance/v1/compliance-rules:
    get:
      description: Fetch compliance rules for UI display.
      operationId: getComplianceRules
      parameters:
      - description: Optional project name filter. The UI may omit this when it only has selected resource ids.
        example: Payment Gateway
        explode: true
        in: query
        name: projectName
        required: false
        schema:
          type: string
        style: form
      - description: Resource type selected in the UI.
        example: MICROSERVICE
        explode: true
        in: query
        name: resourceType
        required: false
        schema:
          $ref: "#/components/schemas/AssetType"
        style: form
      - description: Resource name selected in the UI.
        example: Remit2Any
        explode: true
        in: query
        name: resourceName
        required: false
        schema:
          type: string
        style: form
      - description: Rule status filter.
        example: active
        explode: true
        in: query
        name: status
        required: false
        schema:
          $ref: "#/components/schemas/RuleStatus"
        style: form
      - description: Rule category filter.
        example: DOCUMENTATION
        explode: true
        in: query
        name: ruleCategory
        required: false
        schema:
          $ref: "#/components/schemas/RuleCategory"
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                projectName: Payment Gateway
                resourceType: MICROSERVICE
                resourceName: Remit2Any
                activeRules: 29
                totalRules: 30
                rules:
                - ruleId: CR13
                  ruleName: Documentation required for all endpoints
                  ruleDescription: "Every API endpoint must have comprehensive documentation\
                    \ including request/response schemas, error codes, and usage examples."
                  category: DOCUMENTATION
                  severity: REQUIRED
                  ruleType: PRE_DEFINED
                  ruleOwner: Governance Team
                  enabled: true
                  mandatory: true
                  displayOrder: 1
                  icon: file-text
                  status: active
                  createDate: 2026-05-11T19:30:00Z
                  createdBy: admin@probestack.io
                  updatedDate: 2026-05-11T19:30:00Z
                  updatedBy: admin@probestack.io
              schema:
                $ref: "#/components/schemas/ComplianceRulesResponse"
          description: Compliance rules fetched successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Get compliance rules
      tags:
      - Compliance Rules
      x-accepts: application/json
      x-tags:
      - tag: Compliance Rules
    post:
      description: |
        Adds a new compliance rule.

        ruleId is generated by Java service and should not be sent by UI.
      operationId: createComplianceRule
      requestBody:
        content:
          application/json:
            example:
              ruleName: Documentation required for all endpoints
              ruleDescription: "Every API endpoint must have comprehensive documentation\
                \ including request/response schemas, error codes, and usage examples."
              ruleType: PRE_DEFINED
              ruleOwner: Governance Team
              category: DOCUMENTATION
              severity: REQUIRED
              enabled: true
              mandatory: true
              displayOrder: 1
              icon: file-text
              status: active
              createdBy: admin@probestack.io
              updatedBy: admin@probestack.io
            schema:
              $ref: "#/components/schemas/CreateComplianceRuleRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              example:
                ruleId: CR13
                ruleName: Documentation required for all endpoints
                status: active
                message: Compliance rule created successfully.
              schema:
                $ref: "#/components/schemas/CreateComplianceRuleResponse"
          description: Compliance rule created successfully.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Duplicate rule.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Add compliance rule
      tags:
      - Compliance Rules
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Compliance Rules
  /governance/v1/compliance-rules/{ruleId}/status:
    patch:
      description: Enable, disable, activate, or deactivate a compliance rule from the Governance UI.
      operationId: updateComplianceRuleStatus
      parameters:
      - description: Compliance rule identifier.
        example: CR001
        explode: false
        in: path
        name: ruleId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplianceRuleStatusRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateComplianceRuleStatusResponse"
          description: Compliance rule status updated successfully.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Compliance rule not found.
      summary: Update compliance rule status
      tags:
      - Compliance Rules
...
components:
  schemas:
    AssetType:
      enum:
      - MICROSERVICE
      - APIGEE
      - KONG
      type: string
...
    RuleType:
      enum:
      - PRE_DEFINED
      - CUSTOM
      type: string
    RuleCategory:
      enum:
      - DOCUMENTATION
      - LOGGING
      - SECURITY
      - PERFORMANCE
      - HEALTH
      - GENERAL
      type: string
    RuleSeverity:
      enum:
      - REQUIRED
      - MANDATORY
      - RECOMMENDED
      - OPTIONAL
      type: string
    RuleStatus:
      enum:
      - active
      - inactive
      type: string
...
    CreateComplianceRuleRequest:
      example:
        severity: null
        updatedBy: admin@probestack.io
        displayOrder: 1
        icon: file-text
        mandatory: true
        ruleDescription: "Every API endpoint must have comprehensive documentation\
          \ including request/response schemas, error codes, and usage examples."
        enabled: true
        createdBy: admin@probestack.io
        ruleType: null
        ruleName: Documentation required for all endpoints
        ruleOwner: Governance Team
        category: null
        status: null
      properties:
        ruleName:
          example: Documentation required for all endpoints
          type: string
        ruleDescription:
          example: "Every API endpoint must have comprehensive documentation including\
            \ request/response schemas, error codes, and usage examples."
          type: string
        ruleType:
          $ref: "#/components/schemas/RuleType"
        ruleOwner:
          example: Governance Team
          type: string
        category:
          $ref: "#/components/schemas/RuleCategory"
        severity:
          $ref: "#/components/schemas/RuleSeverity"
        enabled:
          example: true
          type: boolean
        mandatory:
          example: true
          type: boolean
        displayOrder:
          example: 1
          type: integer
        icon:
          example: file-text
          type: string
        status:
          $ref: "#/components/schemas/RuleStatus"
        createdBy:
          example: admin@probestack.io
          type: string
        updatedBy:
          example: admin@probestack.io
          type: string
      required:
      - category
      - createdBy
      - enabled
      - ruleName
      - ruleOwner
      - ruleType
      - severity
      - status
      - updatedBy
      type: object
    CreateComplianceRuleResponse:
      example:
        ruleName: Documentation required for all endpoints
        ruleId: CR13
        message: Compliance rule created successfully.
        status: null
      properties:
        ruleId:
          example: CR13
          type: string
        ruleName:
          example: Documentation required for all endpoints
          type: string
        status:
          $ref: "#/components/schemas/RuleStatus"
        message:
          example: Compliance rule created successfully.
          type: string
      type: object
    ComplianceRulesResponse:
      example:
        resourceName: Remit2Any
        rules:
        - severity: null
          updatedBy: updatedBy
          displayOrder: 0
          icon: icon
          updatedDate: 2000-01-23T04:56:07.000+00:00
          mandatory: true
          ruleDescription: ruleDescription
          enabled: true
          createdBy: createdBy
          ruleType: null
          ruleName: Documentation required for all endpoints
          ruleOwner: Governance Team
          ruleId: CR13
          category: null
          status: null
          createDate: 2000-01-23T04:56:07.000+00:00
        - severity: null
          updatedBy: updatedBy
          displayOrder: 0
          icon: icon
          updatedDate: 2000-01-23T04:56:07.000+00:00
          mandatory: true
          ruleDescription: ruleDescription
          enabled: true
          createdBy: createdBy
          ruleType: null
          ruleName: Documentation required for all endpoints
          ruleOwner: Governance Team
          ruleId: CR13
          category: null
          status: null
          createDate: 2000-01-23T04:56:07.000+00:00
        projectName: Payment Gateway
        activeRules: 29
        totalRules: 30
        resourceType: null
      properties:
        projectName:
          example: Payment Gateway
          type: string
        resourceType:
          $ref: "#/components/schemas/AssetType"
        resourceName:
          example: Remit2Any
          type: string
        activeRules:
          example: 29
          type: integer
        totalRules:
          example: 30
          type: integer
        rules:
          items:
            $ref: "#/components/schemas/ComplianceRule"
          type: array
      type: object
    ComplianceRule:
      example:
        severity: null
        updatedBy: updatedBy
        displayOrder: 0
        icon: icon
        updatedDate: 2000-01-23T04:56:07.000+00:00
        mandatory: true
        ruleDescription: ruleDescription
        enabled: true
        createdBy: createdBy
        ruleType: null
        ruleName: Documentation required for all endpoints
        ruleOwner: Governance Team
        ruleId: CR13
        category: null
        status: null
        createDate: 2000-01-23T04:56:07.000+00:00
      properties:
        ruleId:
          example: CR13
          type: string
        ruleName:
          example: Documentation required for all endpoints
          type: string
        ruleDescription:
          type: string
        category:
          $ref: "#/components/schemas/RuleCategory"
        severity:
          $ref: "#/components/schemas/RuleSeverity"
        ruleType:
          $ref: "#/components/schemas/RuleType"
        ruleOwner:
          example: Governance Team
          type: string
        enabled:
          type: boolean
        mandatory:
          type: boolean
        displayOrder:
          type: integer
        icon:
          type: string
        status:
          $ref: "#/components/schemas/RuleStatus"
        createDate:
          format: date-time
          type: string
        createdBy:
          type: string
        updatedDate:
          format: date-time
          type: string
        updatedBy:
          type: string
      type: object
...
    ErrorResponse:
      properties:
        timestamp:
          format: date-time
          type: string
        status:
          example: 400
          type: integer
        errorCode:
          example: COMPLIANCE_VALIDATION_ERROR
          type: string
        message:
          example: Invalid request payload.
          type: string
        path:
          example: /governance/v1/compliance-rules
          type: string
        serviceTransactionId:
          example: 3d983a40-fbb6-4df0-b43e-34e1c8f8f4bc
          type: string
      type: object
    UpdateComplianceRuleStatusRequest:
      properties:
        enabled:
          example: true
          type: boolean
        status:
          $ref: "#/components/schemas/RuleStatus"
        updatedBy:
          example: admin@probestack.io
          type: string
      type: object
    UpdateComplianceRuleStatusResponse:
      properties:
        ruleId:
          example: CR001
          type: string
        ruleName:
          example: Documentation required for all endpoints
          type: string
        enabled:
          example: true
          type: boolean
        status:
          $ref: "#/components/schemas/RuleStatus"
        message:
          example: Compliance rule status updated successfully.
          type: string
      type: object